Separating Home Scree, Minor bugs. Default Dark theme.
